refactor: standardize Fastly plugin config loading and fix cache inconsistency
- Created dedicated config package (config/fastlyconfig.go) with
  comprehensive test coverage for config validation
- Migrated to commonconfig.GetConfigFilePath() for consistent config
  loading across plugins, following mongodb-atlas plugin pattern
- Fixed cache key format inconsistency: standardized to day-level
  granularity (YYYY-MM-DD_YYYY-MM-DD) for both read and write operations
- Added build tags (//go:build ignore) to utility scripts to prevent
  compilation conflicts with multiple main functions
- Updated go.mod with replace directive for local common module dependency

These changes improve plugin consistency with the opencost plugin system,
eliminate cache misses due to format mismatch, and ensure proper module
dependency resolution.
