+++ pkg/plugins/fastly/utils/create_test_files.go
@@ -1,3 +1,5 @@
+//go:build ignore
+
 package main
 
 import (
