+++ pkg/plugins/fastly/utils/generate_test_data.go
@@ -1,3 +1,5 @@
+//go:build ignore
+
 package main
 
 import (
